Math Puzzle Interaction Working

+++ void_rush/src/common/Helper.cpp
@@ -2,6 +2,6 @@
 
 float deg_to_rad (float deg)
 {
-    float ret = deg * (M_PI / 180);
+    float ret = deg * ((float)M_PI / 180.0f);
     return ret;
 }
