+++ void_rush/src/generation/proto.hpp
@@ -7,7 +7,7 @@
 void generation_proto ()
 {
 
-    int seed = time (0);
+    int seed = (int)time(0);
     std::cout << "Platform generation test\n Input seed: ";
     std::cin >> seed;
     player *pl = new player ();
@@ -40,8 +40,8 @@ void generation_proto ()
         // abs_X += platforms[i]->getPos().at(0) ;
         // abs_Y += platforms[i]->getPos().at(1) ;
 
-        abs_X = platforms[i]->getPos ().y;
-        abs_Y = platforms[i]->getPos ().x;
+        abs_X = (int)platforms[i]->getPos ().y;
+        abs_Y = (int)platforms[i]->getPos ().x;
 
         out << "Platform_" << i << " [\n";
         out << "label = \"P_" << i << "\n";
